- Move utility functions to utils - 64 bytes per line for pem formatting
