+++ packages/sdk/utils/index.ts
@@ -1,4 +1,7 @@
 export * from './base64'
+export * from './bigint'
+export * from './crypto'
 export * from './fetch'
 export * from './nonce'
+export * from './string'
 export * from './url'
